Create the CurrentServerGlobalSetting middleware
